refactor(test-helpers): rename define-schema.ts to schema-types.ts
The defineSchema DSL was deleted in PR #4587, leaving define-schema.ts
holding only the residual schema-type vocabulary (ColumnSpec/Schema/
IndexSpec, COLUMN_TYPE_MAP_*, serialIdType, columnsOf, etc.) consumed by
canonical-schema.ts and TEST_SCHEMA importers. Rename the file to reflect
that residual role and refresh the stale define-schema.ts prose references
across the tree. Test names are left untouched.
