+++ packages/activerecord/src/test-helpers/canonical-schema.ts
@@ -19,14 +19,14 @@ import type {
   AddIndexOptions,
 } from "../connection-adapters/abstract/schema-definitions.js";
 import type { SchemaStatements } from "../connection-adapters/abstract/schema-statements.js";
-import type { ColumnSpec } from "./define-schema.js";
+import type { ColumnSpec } from "./schema-types.js";
 import {
   COLUMN_TYPE_MAP_MYSQL,
   COLUMN_TYPE_MAP_PG,
   COLUMN_TYPE_MAP_SQLITE,
   serialIdType,
   supportsExpressionIndex,
-} from "./define-schema.js";
+} from "./schema-types.js";
 
 interface ColOpts {
   limit?: number;
@@ -2118,7 +2118,7 @@ export async function rebuildCanonicalTables(
   const unknown = [...wanted].filter((n) => !known.has(n));
   if (unknown.length > 0) {
     // Test-helper invariant with no Rails error counterpart — a bare Error is
-    // intentional (mirrors the throws in the sibling define-schema.ts helper).
+    // intentional (mirrors the throws in the sibling schema-types.ts helper).
     // eslint-disable-next-line blazetrails/rails-error-parity
     throw new Error(`rebuildCanonicalTables: unknown canonical table(s): ${unknown.join(", ")}`);
   }
